Improve and refactor
